svc/beancleaner/requirements.txt
aiohttp==3.9.5
aioredis==2.0.1
alpaca-py==0.17.0
apscheduler==3.10.4
redis==5.0.4
